refactory: move as chamadas a populate database/Populate/populate.php para database/Populate/PopulateTrait.php e aplica esse trait em core/Database/Database.php
