+++ database/Populate/populate.php
@@ -3,20 +3,7 @@
 require __DIR__ . '/../../config/bootstrap.php';
 
 use Core\Database\Database;
-use Database\Populate\UserPopulate;
-use Database\Populate\UserRulePopulate;
-use Database\Populate\AccountRulePopulate;
-use Database\Populate\BathroomItemPopulate;
-use Database\Populate\BathroomItemTypePopulate;
-use Database\Populate\BathroomPopulate;
-use Database\Populate\BuildingPopulate;
 
 Database::create();
 Database::migrate();
-UserPopulate::populate();
-UserRulePopulate::populate();
-AccountRulePopulate::populate();
-BuildingPopulate::populate();
-BathroomPopulate::populate();
-BathroomItemTypePopulate::populate();
-BathroomItemPopulate::populate();
+Database::populate();
